# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/search/isrc/{isrc}": {
            "get": {
                "description": "Get track information by ISRC",
                "produces": [
                    "application/json"
                ],
                "summary": "Search tracks by ISRC",
                "operationId": "search-by-isrc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISRC code of the track",
                        "name": "isrc",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Track"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/search/{name}": {
            "get": {
                "description": "Get tracks by artist name",
                "produces": [
                    "application/json"
                ],
                "summary": "Search tracks by artist",
                "operationId": "search-by-artist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Artist"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Artist": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tracks": {
                    "description": "One-to-many relationship with Track",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Track"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Track": {
            "type": "object",
            "properties": {
                "artist": {
                    "description": "Relationship with Artist (parent)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Artist"
                        }
                    ]
                },
                "artistID": {
                    "description": "Foreign key referencing Artist's ID",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "imageURI": {
                    "type": "string"
                },
                "isrc": {
                    "type": "string"
                },
                "popularity": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}